requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-filter>=23.5
django-cors-headers>=4.3.1
python-dotenv>=1.0.0
Pillow>=10.1.0
drf-yasg>=1.21.7 